# secure-design/web-services/api-security-basics/api-specifications/OpenAPI-Generator/src/main/resources/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: 1.0.0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /v1/articles/{id}:
    get:
      security:
        - basicAuth: []
      tags:
      - article-controller
      operationId: one
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            'application/json':
              schema:
                type: string
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Article'
    put:
      security:
        - basicAuth: []
      tags:
      - article-controller
      operationId: replaceEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            'application/json':
              schema:
                type: string
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Article'
    delete:
      security:
        - basicAuth: [ ]
      tags:
      - article-controller
      operationId: deleteEmployee
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
          content:
            'application/json':
              schema:
                type: string
        "200":
          description: OK
  /v1/articles:
    get:
      security:
        - basicAuth: []
      tags:
      - article-controller
      summary: Get all Articles
      description: Returns a list of articles
      operationId: articles
      responses:
        "401":
          description: Unauthorized
        "404":
          description: Not found
          content:
            'application/json':
              schema:
                type: string
        "200":
          description: "ok, successful operation"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
    post:
      security:
        - basicAuth: []
      tags:
      - article-controller
      operationId: newArticle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            'application/json':
              schema:
                type: string
        "200":
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/Article'

components:
  securitySchemes:
    basicAuth:     
      type: http
      scheme: basic

  schemas:
    Article:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
          maxLength: 1024
        price:
          type: integer
          format: int64
          minimum: 0